Extract common utils.
